Move external out of src
