+++ example/demo_cli.dart
@@ -3,7 +3,7 @@
 import 'dart:io';
 
 import 'package:dcli/dcli.dart';
-import 'package:trello_sdk/src/sdk/external/dio_client_factory.dart';
+import 'package:trello_sdk/external/dio_client_factory.dart';
 import 'package:trello_sdk/trello_cli.dart';
 
 const String MEMBER = "kemitix";
