+++ lib/external/files.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import '../cards/card_models.dart';
+import '../src/sdk/cards/card_models.dart';
 
 Future<void> defaultFileWriter(FileName fileName, dynamic data) {
   var f = File(fileName.value).openSync(mode: FileMode.write);
